backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
sqlalchemy
spacy
mediapipe
opencv-python
openai-whisper
python-dotenv
